+++ src/App/Shared/GoldEx.Shared/DTOs/Customers/GetCustomerResponse.cs
@@ -8,6 +8,6 @@ public record GetCustomerResponse(
     string NationalId,
     string PhoneNumber,
     string? Address,
-    double? CreditLimit,
+    decimal? CreditLimit,
     UnitType? CreditLimitUnit,
     CustomerType CustomerType);
