+++ src/App/Shared/GoldEx.Shared/DTOs/Settings/CreateSettingRequest.cs
@@ -4,6 +4,6 @@ public record CreateSettingRequest(
     string InstitutionName,
     string Address,
     string PhoneNumber,
-    float TaxPercent,
-    float GoldProfitPercent,
-    float JewelryProfitPercent);
+    decimal TaxPercent,
+    decimal GoldProfitPercent,
+    decimal JewelryProfitPercent);
