+++ src/App/Shared/GoldEx.Shared/Settings/DefaultSetting.cs
@@ -5,7 +5,7 @@ public class DefaultSetting
     public string InstitutionName { get; set; } = default!;
     public string Address { get; set; } = "تهران، خیابان ولیعصر، میدان ولیعصر، پلاک 1";
     public string PhoneNumber { get; set; } = "09905492104";
-    public float TaxPercent { get; set; }
-    public float GoldProfitPercent { get; set; }
-    public float JewelryProfitPercent { get; set; }
+    public decimal TaxPercent { get; set; }
+    public decimal GoldProfitPercent { get; set; }
+    public decimal JewelryProfitPercent { get; set; }
 }
